chore: extract link_endpoint_oauth to keep handlers.rs under 1000-line cap
PR #195 fix-batch commit (571ca51) added ~50 lines to handlers.rs which
pushed it from 957 to 1026 lines. CI file-size gate failed (1000-line cap
per project CLAUDE.md). Move LinkOauthBody + link_endpoint_oauth to new
crates/rs-api/src/endpoint_oauth.rs. handlers.rs back to 957 lines.
